requirements.txt
streamlit
pdfplumber
python-docx
langdetect
